// src/components/Github-Profile-Wrapper/types.ts
export interface GithubProfile {
    login: string;
    avatar_url: string;
    html_url: string;
    bio: string;
    created_at: string;
    followers: number;
    following: number;
    name: string;
    public_repos: number;
    repos_url: string;
    updated_at: string;
    message?: string; 
  }
  